+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - programs(store): Added `close_empty_position` instruction.
 - programs(store): Added flag to indicate whether the token's market is open in oracle price map.
 - programs(store): Added `update_closed_state` instruction.
+- programs(store): Added new market config: `min_collatereal_factor_for_liquidation`.
 - sdk(sdk): Added `try_deserialize_zero_copy_from_base64_with_options` utility function.
 - sdk(js): Added decode methods with `no_discriminator` option for JsMarket and JsPosition.
 - sdk(js): Added `add_create_orders_builder` function.